add core files in 'app' folder
+ refactored data_search
+ follow the new skeleton now :)
