+++ app/data_connexion.py
@@ -3,3 +3,4 @@ DATABASE_NAME = 'openfoodfact'
 HOST = '127.0.0.1'
 USER = 'root'
 PASSWORD = '4405'
+
