+++ app/openFOODfacts.py
@@ -3,7 +3,7 @@
 class Category():
     def __init__(self, name):
         self.name = name
-        
+
 
 class Product():
     def __init__(self, nutriscore_id, name, code, description, url, store):
